moved connected input into common
