+++ src/renderer/hooks/useValidity.ts
@@ -3,9 +3,9 @@ import { useReactFlow } from 'reactflow';
 import { useContextSelector } from 'use-context-selector';
 import { EdgeData, InputData, NodeData, NodeSchema } from '../../common/common-types';
 import { checkNodeValidity, checkRequiredInputs } from '../../common/nodes/checkNodeValidity';
+import { getConnectedInputs } from '../../common/nodes/connectedInputs';
 import { VALID, Validity, invalid } from '../../common/Validity';
 import { GlobalVolatileContext } from '../contexts/GlobalNodeState';
-import { getConnectedInputs } from '../helpers/connectedInputs';
 
 const STARTING_VALIDITY: Validity = invalid('Validating nodes...');
